refactor: extract/consolidate indexing capabilities to a separate component
