+++ redis-om-spring/src/main/java/com/redis/om/spring/repository/support/RedisEnhancedRepositoryFactoryBean.java
@@ -13,7 +13,7 @@ import org.springframework.data.repository.query.parser.AbstractQueryCreator;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
 
-import com.redis.om.spring.KeyspaceToIndexMap;
+import com.redis.om.spring.RediSearchIndexer;
 import com.redis.om.spring.ops.RedisModulesOperations;
 import com.redis.om.spring.repository.query.RedisEnhancedQuery;
 
@@ -23,7 +23,7 @@ public class RedisEnhancedRepositoryFactoryBean<T extends Repository<S, ID>, S, 
   private @Nullable KeyValueOperations operations;
   private @Nullable RedisModulesOperations<String> rmo;
   private @Nullable RedisOperations<?, ?> redisOperations;
-  private @Nullable KeyspaceToIndexMap keyspaceToIndexMap;
+  private @Nullable RediSearchIndexer indexer;
   private @Nullable Class<? extends AbstractQueryCreator<?, ?>> queryCreator;
   private @Nullable Class<? extends RepositoryQuery> repositoryQueryType;
 
@@ -37,7 +37,7 @@ public class RedisEnhancedRepositoryFactoryBean<T extends Repository<S, ID>, S, 
    * @param keyspaceToIndexMap must not be {@literal null}.
    */
   public RedisEnhancedRepositoryFactoryBean(Class<? extends T> repositoryInterface,
-      RedisOperations<?, ?> redisOperations, RedisModulesOperations<?> rmo, KeyspaceToIndexMap keyspaceToIndexMap) {
+      RedisOperations<?, ?> redisOperations, RedisModulesOperations<?> rmo, RediSearchIndexer keyspaceToIndexMap) {
     super(repositoryInterface);
     setRedisModulesOperations(rmo);
     setRedisOperations(redisOperations);
@@ -90,8 +90,8 @@ public class RedisEnhancedRepositoryFactoryBean<T extends Repository<S, ID>, S, 
     super.setMappingContext(mappingContext);
   }
   
-  public void setKeyspaceToIndexMap(KeyspaceToIndexMap keyspaceToIndexMap) { 
-    this.keyspaceToIndexMap = keyspaceToIndexMap;
+  public void setKeyspaceToIndexMap(RediSearchIndexer keyspaceToIndexMap) { 
+    this.indexer = keyspaceToIndexMap;
   }
 
   /**
@@ -137,7 +137,7 @@ public class RedisEnhancedRepositoryFactoryBean<T extends Repository<S, ID>, S, 
    */
   protected RedisEnhancedRepositoryFactory createRepositoryFactory(KeyValueOperations operations,
       Class<? extends AbstractQueryCreator<?, ?>> queryCreator, Class<? extends RepositoryQuery> repositoryQueryType) {
-    return new RedisEnhancedRepositoryFactory(operations, redisOperations, rmo, keyspaceToIndexMap, queryCreator, RedisEnhancedQuery.class);
+    return new RedisEnhancedRepositoryFactory(operations, redisOperations, rmo, indexer, queryCreator, RedisEnhancedQuery.class);
   }
 
   /* (non-Javadoc)
@@ -152,7 +152,7 @@ public class RedisEnhancedRepositoryFactoryBean<T extends Repository<S, ID>, S, 
     Assert.notNull(rmo, "RedisModulesOperations must not be null!");
     Assert.notNull(queryCreator, "Query creator must not be null!");
     Assert.notNull(repositoryQueryType, "RepositoryQueryType must not be null!");
-    Assert.notNull(keyspaceToIndexMap, "KeyspaceToIndexMap type must not be null");
+    Assert.notNull(indexer, "KeyspaceToIndexMap type must not be null");
 
     super.afterPropertiesSet();
   }
