+++ redis-om-spring/src/main/java/com/redis/om/spring/util/ObjectUtils.java
@@ -7,12 +7,18 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider;
 import org.springframework.core.ResolvableType;
+import org.springframework.core.type.filter.AnnotationTypeFilter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.geo.Distance;
 import org.springframework.data.redis.connection.RedisGeoCommands.DistanceUnit;
@@ -20,7 +26,11 @@ import org.springframework.data.redis.core.convert.Bucket;
 import org.springframework.data.redis.core.convert.RedisData;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 import org.springframework.util.ReflectionUtils;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import com.redis.om.spring.annotations.EnableRedisDocumentRepositories;
+import com.redis.om.spring.annotations.EnableRedisEnhancedRepositories;
 import com.redis.om.spring.convert.MappingRedisOMConverter;
 
 import io.redisearch.Document;
@@ -247,6 +257,50 @@ public class ObjectUtils {
   public static String asString(Object value, MappingRedisOMConverter mappingConverter) {
     return value instanceof String ? (String) value
         : mappingConverter.getConversionService().convert(value, String.class);
+  }
+  
+  @SuppressWarnings({ "unchecked", "rawtypes" })
+  public static Set<BeanDefinition> getBeanDefinitionsFor(ApplicationContext ac, Class... classes) {
+    Map<String, Object> annotatedBeans = ac.getBeansWithAnnotation(SpringBootApplication.class);
+    Class<?> app = annotatedBeans.values().toArray()[0].getClass();
+    Set<BeanDefinition> beanDefs = new HashSet<>();
+
+    ClassPathScanningCandidateComponentProvider provider = new ClassPathScanningCandidateComponentProvider(false);
+    for (Class cls : classes) {
+      provider.addIncludeFilter(new AnnotationTypeFilter(cls));
+    }
+
+    if (app.isAnnotationPresent(EnableRedisDocumentRepositories.class)) {
+      EnableRedisDocumentRepositories edr = app.getAnnotation(EnableRedisDocumentRepositories.class);
+      if (edr.basePackages().length > 0) {
+        for (String pkg : edr.basePackages()) {
+          beanDefs.addAll(provider.findCandidateComponents(pkg));
+        }
+      } else if (edr.basePackageClasses().length > 0) {
+        for (Class<?> pkg : edr.basePackageClasses()) {
+          beanDefs.addAll(provider.findCandidateComponents(pkg.getPackageName()));
+        }
+      } else {
+        beanDefs.addAll(provider.findCandidateComponents(app.getPackageName()));
+      }
+    }
+
+    if (app.isAnnotationPresent(EnableRedisEnhancedRepositories.class)) {
+      EnableRedisEnhancedRepositories er = app.getAnnotation(EnableRedisEnhancedRepositories.class);
+      if (er.basePackages().length > 0) {
+        for (String pkg : er.basePackages()) {
+          beanDefs.addAll(provider.findCandidateComponents(pkg));
+        }
+      } else if (er.basePackageClasses().length > 0) {
+        for (Class<?> pkg : er.basePackageClasses()) {
+          beanDefs.addAll(provider.findCandidateComponents(pkg.getPackageName()));
+        }
+      } else {
+        beanDefs.addAll(provider.findCandidateComponents(app.getPackageName()));
+      }
+    }
+
+    return beanDefs;
   }
 
   private ObjectUtils() {}
